backend/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
caer==2.0.8
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
Flask==3.1.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meson==1.6.0
motor==3.6.0
mypy==1.13.0
mypy-extensions==1.0.0
narwhals==1.15.2
ninja==1.11.1.2
numpy==2.1.3
opencv-contrib-python==4.10.0.84
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
protobuf==5.29.1
pyarrow==18.1.0
pyasn1==0.6.1
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.40.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
utils==1.0.2
uvicorn==0.32.1
watchdog==6.0.0
Werkzeug==3.1.3
